Updated BigDecimal to Double as Firebase does not support BigDecimal

+++ src/main/java/com/subtracker/model/Subscription.java
@@ -4,8 +4,8 @@ import com.google.firebase.database.annotations.NotNull;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.Date;
+
 @Data
 @NoArgsConstructor
 public class Subscription {
@@ -28,8 +28,8 @@ public class Subscription {
     private String currency;
 
     @NotNull
-    private BigDecimal monthlyAmount;
-    private BigDecimal yearlyAmount;
+    private Double monthlyAmount;
+    private Double yearlyAmount;
 
     // In months
     private double contractualPeriod;
